Refactor midonet-cli to use the new style classes
`super()` is only supported in Python3 and I just put the class name and
`self` as an instance for the arguments of `__init__`. It's not
abstracted well but the follwing idiom doesn't work for the classes
which have further derived class because `self.__class__` would be
always the class of the instance `self` and cause the stack overflow.

    def __init__(self, obj):
        super(self.__class__, self).__init__(obj)

See the following links for the further information:

  * http://stackoverflow.com/questions/54867/
  * http://stackoverflow.com/questions/576169/
  * http://www.artima.com/weblogs/viewpost.jsp?thread=236275

Signed-off-by: Taku Fukushima <tfukushima@midokura.com>